+++ app/src/main/java/com/perol/asdpl/pixivez/ui/pic/PicturePagerAdapter.kt
@@ -31,7 +31,7 @@ import com.perol.asdpl.pixivez.data.model.Illust
 
 class PicturePagerAdapter(
     fm: FragmentManager,
-    private val idList: LongArray?,
+    private val idList: IntArray?,
     private val illustList: MutableList<Illust>? = null
 ) : FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
     override fun getItem(position: Int) = PictureXFragment.newInstance(
